pyproject.toml
[project]
name = "test-work"
version = "0.1.0"
description = ""
authors = [
    {name = "Valera Zaycev",email = "admin@frofy.ru"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all] (>=0.115.12,<0.116.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pydantic (>=2.11.2,<3.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "loguru (==0.7.2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
py-dotenv = "^0.1"

